backend/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5

# Google Cloud Services
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.34.0
google-cloud-firestore==2.11.1
firebase-admin==6.2.0

# AI and Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3

# Security and Encryption
cryptography==41.0.3
bcrypt==4.0.1
PyJWT==2.8.0

# File Processing
Pillow==10.0.0
python-magic==0.4.27
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.1
redis==4.6.0

# Development and Testing
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
